requirements.txt
arrow==0.14.5
astroid==2.2.5
bitarray==1.0.1
blessed==1.15.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-picklefield==2.0
django-q==1.0.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.3
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
django-model-utils==3.2.0
mock==3.0.5
fyle-django-allauth==0.1.0
pyxero==0.9.1
django-decorator-include==2.1
